move assembleUrl back now that it's only being used by torznab
